# docs/architecture/openapi.yaml
openapi: 3.0.3
info:
  title: beelms API
  version: 0.1.0
  license:
    name: UNLICENSED
    url: https://example.com/unlicense
  description: |
    OpenAPI спецификация за основните MVP услуги на beelms.
servers:
  - url: https://api.beelms.com/api
    description: Production (примерен домейн, с /api префикс)
  - url: http://127.0.0.1:3000/api
    description: Local development (с /api префикс)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties: true
...
    PublicSettings:
      type: object
      properties:
        branding:
          type: object
          properties:
            appName:
              type: string
            logoUrl:
              type: string
              nullable: true
            primaryColor:
              type: string
              nullable: true
          required: [appName]
        features:
          type: object
          properties:
            wikiPublic:
              type: boolean
            courses:
              type: boolean
            auth:
              type: boolean
            paidCourses:
              type: boolean
            gdprLegal:
              type: boolean
            infraRedis:
              type: boolean
            infraRabbitmq:
              type: boolean
            infraMonitoring:
              type: boolean
            infraErrorTracking:
              type: boolean
          required:
            [wikiPublic, courses, auth, paidCourses, gdprLegal, infraRedis, infraRabbitmq, infraMonitoring, infraErrorTracking]
        languages:
          type: object
          properties:
            supported:
              type: array
              items:
                type: string
            default:
              type: string
          required: [supported, default]
      required: [branding, features, languages]
...
    AdminUpdateInstanceSettingsInput:
      type: object
      properties:
        branding:
          type: object
          properties:
            appName:
              type: string
            logoUrl:
              type: string
              nullable: true
            primaryColor:
              type: string
              nullable: true
        features:
          type: object
          properties:
            wikiPublic:
              type: boolean
            courses:
              type: boolean
            auth:
              type: boolean
            paidCourses:
              type: boolean
            gdprLegal:
              type: boolean
            infraRedis:
              type: boolean
            infraRabbitmq:
              type: boolean
            infraMonitoring:
              type: boolean
            infraErrorTracking:
              type: boolean
        languages:
          type: object
          properties:
            supported:
              type: array
              items:
                type: string
            default:
              type: string
...
    WikiArticle:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        language:
          type: string
          example: bg
        title:
          type: string
        content:
          type: string
        status:
          type: string
          description: Статус на статията (draft, active, inactive)
          enum: [draft, active, inactive]
        updatedAt:
          type: string
          format: date-time
...
    CourseSummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        language:
          type: string
          example: bg
        status:
          type: string
          description: Статус на курса (draft, active, inactive)
          enum: [draft, active, inactive]
        isPaid:
          type: boolean
          description: Дали курсът е платен (изисква плащане преди enrollment)
        currency:
          type: string
          nullable: true
        priceCents:
          type: integer
          nullable: true
        categoryId:
          type: string
          nullable: true
        category:
          oneOf:
            - $ref: '#/components/schemas/CourseCategorySummary'
            - type: object
              nullable: true
    CourseCategorySummary:
      type: object
      properties:
        slug:
          type: string
        title:
          type: string
...
    CourseModuleItem:
      type: object
      properties:
        id:
          type: string
        itemType:
          type: string
          description: Тип на елемента в програмата на курса
          enum: [wiki, task, quiz]
        title:
          type: string
        order:
          type: integer
        wikiSlug:
          type: string
          nullable: true
        taskId:
          type: string
          nullable: true
        quizId:
          type: string
          nullable: true
    CourseDetail:
      allOf:
        - $ref: '#/components/schemas/CourseSummary'
        - type: object
          properties:
            curriculum:
              type: array
              items:
                $ref: '#/components/schemas/CourseModuleItem'
...
    CourseTaskDetail:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        language:
          type: string
        status:
          type: string
        updatedAt:
          type: string
          format: date-time
    CourseCertificate:
      type: object
      properties:
        courseId:
          type: string
        courseTitle:
          type: string
        userId:
          type: string
        userEmail:
          type: string
        completedAt:
          type: string
          format: date-time
        issuedAt:
          type: string
          format: date-time
...
    AdminPaymentSettings:
      type: object
      properties:
        currency:
          type: string
        priceCents:
          type: integer
      required: [currency, priceCents]
    AdminUpdatePaymentSettingsInput:
      type: object
      properties:
        currency:
          type: string
          nullable: true
        priceCents:
          type: integer
          nullable: true
    AdminStripeWebhookEvent:
      type: object
      properties:
        id:
          type: string
        eventId:
          type: string
        eventType:
          type: string
        status:
          type: string
          enum: [received, processed, failed]
        errorMessage:
          type: string
          nullable: true
        processedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AdminRetryWebhookEventResponse:
      type: object
      properties:
        status:
          type: string
          enum: [processed, failed]
        errorMessage:
          type: string
          nullable: true
...
security:
  - bearerAuth: []
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [Courses]
      summary: Списък с курсове (catalog)
      security: []
      parameters:
        - in: query
          name: category
          required: false
          schema:
            type: string
          description: Филтър по category slug
      responses:
        "200":
          description: Списък с курсове
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseSummary'
        "400":
          description: Невалидни параметри
...
  /courses/{courseId}:
    get:
      operationId: getCourse
      tags: [Courses]
      summary: Детайл на курс
      security: []
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Детайл на курс
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDetail'
        "404":
          description: Курсът не е намерен
  /courses/{courseId}/enroll:
    post:
      operationId: enrollInCourse
      tags: [Courses]
      summary: Записване (enroll) в курс
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Потребителят е записан успешно
        "401":
          description: Неоторизиран достъп
        "403":
          description: Payment required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Курсът не е намерен
...
  /courses/{courseId}/tasks/{taskId}:
    get:
      operationId: getCourseTask
      tags: [Courses]
      summary: Детайл на task (в рамките на курс)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: string
        - in: path
          name: taskId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseTaskDetail'
        "401":
          description: Неоторизиран достъп
        "403":
          description: Няма достъп (не е записан / няма purchase)
        "404":
          description: Task-ът не е намерен
  /courses/{courseId}/wiki/{slug}:
    get:
      operationId: getCourseWikiArticle
      tags: [Courses, Wiki]
      summary: Детайл на wiki статия в рамките на курс
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: string
        - in: path
          name: slug
          required: true
          schema:
            type: string
        - in: query
          name: lang
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Wiki article detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WikiArticle'
        "401":
          description: Неоторизиран достъп
        "403":
          description: Няма достъп (не е записан / няма purchase)
        "404":
          description: Статията не е намерена
  /courses/{courseId}/certificate:
    get:
      operationId: getCourseCertificate
      tags: [Courses]
      summary: Certificate за завършен курс
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Course certificate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseCertificate'
        "401":
          description: Неоторизиран достъп
        "403":
          description: Course not completed
...
  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [Admin]
      summary: Current instance settings (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current instance settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicSettings'
        "401":
          description: Неоторизиран достъп
        "403":
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      operationId: updateAdminSettings
      tags: [Admin]
      summary: Update instance settings (admin)
      description: Partial update (PATCH) for branding/features/languages.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateInstanceSettingsInput'
      responses:
        "200":
          description: Updated instance settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicSettings'
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Неоторизиран достъп
        "403":
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/payments/currencies:
    get:
      operationId: listAdminPaymentCurrencies
      tags: [Admin]
      summary: Supported payment currencies
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Supported currencies
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "401":
          description: Неоторизиран достъп
  /admin/payments/settings:
    get:
      operationId: getAdminPaymentSettings
      tags: [Admin]
      summary: Payment settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Payment settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPaymentSettings'
        "401":
          description: Неоторизиран достъп

    patch:
      operationId: updateAdminPaymentSettings
      tags: [Admin]
      summary: Update payment settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdatePaymentSettingsInput'
      responses:
        "204":
          description: Settings updated
        "400":
          description: Невалидни данни
        "401":
          description: Неоторизиран достъп
  /admin/payments/webhook-events:
    get:
      operationId: listAdminPaymentWebhookEvents
      tags: [Admin]
      summary: List Stripe webhook events
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [received, processed, failed]
        - in: query
          name: limit
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Webhook events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminStripeWebhookEvent'
        "400":
          description: Невалидни данни
        "401":
          description: Неоторизиран достъп
  /admin/payments/webhook-events/{eventId}/retry:
    post:
      operationId: retryAdminPaymentWebhookEvent
      tags: [Admin]
      summary: Retry processing of a webhook event
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Retry result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminRetryWebhookEventResponse'
        "401":
          description: Неоторизиран достъп

  ############################
  # Admin Courses (CRUD)
  ############################
...
  /admin/courses/{courseId}:
    get:
      operationId: getAdminCourse
      tags: [Admin, Courses]
      summary: Детайл на курс за администрация
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Детайл на курс
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDetail'
        "401":
          description: Неоторизиран достъп
        "404":
          description: Курсът не е намерен

    patch:
      operationId: updateAdminCourse
      tags: [Admin, Courses]
      summary: Редакция на курс
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                language:
                  type: string
                status:
                  type: string
                  enum: [draft, active, inactive]
                isPaid:
                  type: boolean
                currency:
                  type: string
                  nullable: true
                priceCents:
                  type: integer
                  nullable: true
                categoryId:
                  type: string
                  nullable: true
      responses:
        "200":
          description: Обновен курс
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDetail'
        "400":
          description: Невалидни данни
        "401":
          description: Неоторизиран достъп
        "404":
          description: Курсът не е намерен
...